Bug fixing 1.0.1

+++ src/main/java/net/andrespr/casinorocket/games/slot/SlotLineResult.java
@@ -5,6 +5,6 @@ public record SlotLineResult(
         SlotSymbol symbol,
         int count,
         int multiplier,
-        int lineWin,
+        long lineWin,
         int lineIndex
 ) {}
